requirements (2).txt
streamlit>=1.24.0
pandas>=1.5.3
mysql-connector-python>=8.0.32
python-dotenv>=1.0.0
